schemasnapshot: split capture orchestrator into capture.go
Move Capture/CaptureParams/CaptureRequest/CaptureAndSaveSnapshot out of
provider.go into capture.go. provider.go now holds only the provider SPI
(QueryExecutor, SnapshotProvider) and the registry. Pure file move — no
behavior change. The capture orchestrator is the library entry point, so
it reads better in its own file than under the provider registry.

---
_automated · Claude Code (Opus 4.8)_
